Dec 8: 2h, break interpreter into multiple files
